refactor: reorganize backend structure
