--- backend/cmd/server/main.go
@@ -1,65 +1,0 @@
-package main
-
-import (
-	"log"
-	"net/http"
-	"time"
-
-	"github.com/gin-gonic/gin"
-	"go.uber.org/zap"
-
-	"github.com/hungdhv97/english-vocab-trainer/backend/internal/config"
-	"github.com/hungdhv97/english-vocab-trainer/backend/internal/handler"
-	"github.com/hungdhv97/english-vocab-trainer/backend/internal/service"
-)
-
-func main() {
-	cfg, err := config.Load()
-	if err != nil {
-		log.Fatalf("load config: %v", err)
-	}
-
-	logger, _ := zap.NewProduction()
-	defer logger.Sync()
-
-	svc := service.NewService(cfg)
-	hdl := handler.NewHandler(svc)
-
-	r := gin.New()
-	r.Use(gin.Recovery())
-
-	r.Use(func(c *gin.Context) {
-		start := time.Now()
-		c.Next()
-		logger.Info("request",
-			zap.String("method", c.Request.Method),
-			zap.String("path", c.Request.URL.Path),
-			zap.Int("status", c.Writer.Status()),
-			zap.Duration("duration", time.Since(start)))
-	})
-
-	r.Use(func(c *gin.Context) {
-		origin := c.GetHeader("Origin")
-		if origin != "" {
-			c.Writer.Header().Set("Access-Control-Allow-Origin", origin)
-			c.Writer.Header().Set("Vary", "Origin")
-			c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
-		} else {
-			c.Writer.Header().Set("Access-Control-Allow-Origin", "*")
-		}
-		c.Writer.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
-		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
-		if c.Request.Method == http.MethodOptions {
-			c.AbortWithStatus(http.StatusNoContent)
-			return
-		}
-		c.Next()
-	})
-
-	hdl.RegisterRoutes(r)
-
-	logger.Info("Starting server", zap.String("addr", cfg.HTTP.Addr))
-	if err := r.Run(cfg.HTTP.Addr); err != nil {
-		logger.Fatal("server failed", zap.Error(err))
-	}
-}
