--- backend/internal/auth/auth.go
@@ -1,6 +1,0 @@
-package auth
-
-// Authenticator is the authentication layer.
-type Authenticator interface {
-	// Add methods here
-}
