--- backend/internal/db/db.go
@@ -1,6 +1,0 @@
-package db
-
-// DB is the database interface.
-type DB interface {
-	// Add methods here
-}
