+++ backend/internal/modules/user/dto/auth.go
@@ -1,0 +1,13 @@
+package dto
+
+// RegisterRequest represents a registration request.
+type RegisterRequest struct {
+	Username string `json:"username" validate:"required"`
+	Password string `json:"password" validate:"required"`
+}
+
+// LoginRequest represents a login request.
+type LoginRequest struct {
+	Username string `json:"username" validate:"required"`
+	Password string `json:"password" validate:"required"`
+}
