+++ backend/internal/modules/user/handler/http.go
@@ -1,0 +1,60 @@
+package handler
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/go-playground/validator/v10"
+
+	"github.com/hungdhv97/english-vocab-trainer/backend/internal/modules/user/dto"
+	"github.com/hungdhv97/english-vocab-trainer/backend/internal/modules/user/service"
+)
+
+// Handler provides HTTP handlers for user endpoints.
+type Handler struct {
+	svc      *service.Service
+	validate *validator.Validate
+}
+
+// New creates a new user handler.
+func New(s *service.Service) *Handler {
+	return &Handler{svc: s, validate: validator.New()}
+}
+
+// Register handles user registration.
+func (h *Handler) Register(c *gin.Context) {
+	var req dto.RegisterRequest
+	if err := c.ShouldBindJSON(&req); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	if err := h.validate.Struct(req); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	user, err := h.svc.Register(req.Username, req.Password)
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	c.JSON(http.StatusOK, user)
+}
+
+// Login handles user authentication.
+func (h *Handler) Login(c *gin.Context) {
+	var req dto.LoginRequest
+	if err := c.ShouldBindJSON(&req); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	if err := h.validate.Struct(req); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	user, err := h.svc.Authenticate(req.Username, req.Password)
+	if err != nil {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
+		return
+	}
+	c.JSON(http.StatusOK, user)
+}
