+++ backend/internal/platform/config/config.go
@@ -1,0 +1,29 @@
+package config
+
+// Config holds application configuration.
+type Config struct {
+	HTTP     HTTPConfig     `mapstructure:"http"`
+	Postgres PostgresConfig `mapstructure:"postgres"`
+	Redis    RedisConfig    `mapstructure:"redis"`
+}
+
+// HTTPConfig holds HTTP server related configuration.
+type HTTPConfig struct {
+	Addr string `mapstructure:"addr"`
+}
+
+// PostgresConfig holds PostgreSQL connection settings.
+type PostgresConfig struct {
+	Host     string `mapstructure:"host"`
+	Port     string `mapstructure:"port"`
+	User     string `mapstructure:"user"`
+	Password string `mapstructure:"password"`
+	Database string `mapstructure:"database"`
+}
+
+// RedisConfig holds Redis connection settings.
+type RedisConfig struct {
+	Addr     string `mapstructure:"addr"`
+	Username string `mapstructure:"username"`
+	Password string `mapstructure:"password"`
+}
